Fix mkdir_and_cd
Previously the mkdir_and_cd command would not handle more than one
argument. So, you wouldn't be able to send flags to it. This fixes that.

The $@ tells zsh to pass all the arguments to mkdir and ${@[$#]} gets
the last argument which should be the filename.

The parts of this are as follow:
1. $@ - defines the array of arguments
2. $# - get the last element of the array
3. ${@[$#]} - get the last element of the array of arguments.