Add toggleterm
This is a terminal for the vim. It persists between open/close. I also
added some keymaps, but I was having some trouble. I'm not sure if I'll
be able to figure out the keymaps, so I might just use it in float mode
or in vertical mode in the future.