Ignore .folders
These folders are in the tree but aren't useful and clutter the
treeview. This removes them. In the future, I might want to make this
more smart or find a way to hide them by default, but load them when I
need them similar to how Telescope had multiple options