Add vim-pencil to nvim-plugins
vim-pencil provides a bunch of features that make writing prose a lot
better in vim. I've added it as a part of working on the new version of
`eliseshaffer.com`.

I also added this for git commit messages bc I think it will be helpful
with crafting well written commit messages.