Update plugins
Add winbar configuration and updated to the new version of auto-save.
I also tried updating feline, but it ended up not really working with my
colorswitching. I think the issue is that feline might not support
theming but the changelog doesn't say anything about it. I pinned the
tag to v0.4.0 for now and I'll try to dig into it later.