Fix Neovim ruby version error
While running RSpec tests, I noticed that it was using the system ruby
instead of the ruby version from chruby. So I added the correct path in
my zshenv.

This has a drawback in that is means that the envvar is set even after
I call chruby. Which is not ideal.

This might be because I swapped it from a sourced file to a zshenv. I'll
need to investigate this further.