Add ufo and statuscol plugins for Neovim
Ufo helps with intelligent folding and statuscol makes the status column
configurable. I decided to use LSP for the folding login, but I think I
might be better off with Treesitter. I'm gonna try if for now, but might
switch later.

The main issue I've noticed is that it can take a minute for the folds
to load. I'm not sure if thats bc its LSP or just an issue with UFO.