Use dotbot for dotfiles management
dotbot is a dotfiles manager that uses a yaml file to set up symlinks.
I'm trying it because I think it's a lot simpler than my install
scripts. So far this only links my dotfiles.
It doesn't handle the additional things like running `brew bundle` or
installing Ruby or Node. Those will be handled in the future. For now,
I'm keeping my install script for now. We'll see how well this goes.