Add embedded_template treesitter highlighting
This will help fix some highlighting issues with eruby. I also am
treating markdown files as eruby files so I can get decent highlighting
on a few pages. But I might try to separate out the markdown data and
the templates in the future. So that might not be necessary forever